Moved detection idiom to meta and revised all ranges algorithm
